requirements.txt
playwright
python-telegram-bot>=20.0
rich
python-dotenv
rapidfuzz
colorama